Move quantized impls into kernels::quantized
